ML_embeddings/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
joblib>=1.3.0
pyarrow>=14.0.0
tqdm>=4.66.0
scikit-learn>=1.3.0  # Para análisis y métricas avanzadas

# Para modo GRANDE (100k recetas) 
datasets>=2.16.0

